pylon/server/requirements/common.txt
#
# Copyright (c) 2013 Echelon Corporation.  All rights reserved.
# 
# Permission is hereby granted, free of charge, to any person obtaining a copy
# of this software and associated documentation files (the "Software"), to deal
# in the Software without restriction, including without limitation the rights
# to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
# copies of the Software, and to permit persons to whom the Software is
# furnished to do so, subject to the following conditions:
# 
# The above copyright notice and this permission notice shall be included in
# all copies or substantial portions of the Software.
# 
# THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
# IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
# FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
# AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
# LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
# OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
# THE SOFTWARE.
#

# common requirements
#
# install using:
#   pip install -r requirements/dev.txt
# or:
#   pip install -r requirements/prod.txt

# Django
Django>=1.5.2
django-filter>=0.7
#django-oauth2-provider>=0.2.6		# Python 2 only

# Django REST Framework
djangorestframework>=2.3.7
South>=0.8.2
gunicorn>=17.5
Markdown>=2.3.1

# Django REST Framework Parsers
PyYAML>=3.10
defusedxml>=0.4.1

#elementtree				# Python 2 only?
#cElementTree				# Python 2 only?

# IP-C Server
setproctitle>=1.1.7

# IP-C Server Localhost Collector
psutil>=1.0.1
